+++ src/main/java/in/hocg/web/filter/group/Delete.java
@@ -1,0 +1,8 @@
+package in.hocg.web.filter.group;
+
+/**
+ * Created by hocgin on 2017/11/8.
+ * email: hocgin@gmail.com
+ */
+public interface Delete {
+}