+++ src/main/java/in/hocg/web/modules/service/VariableService.java
@@ -1,7 +1,6 @@
 package in.hocg.web.modules.service;
 
-import in.hocg.web.filter.VariableInsertFilter;
-import in.hocg.web.filter.VariableUpdateFilter;
+import in.hocg.web.filter.VariableFilter;
 import in.hocg.web.lang.CheckError;
 import in.hocg.web.modules.domain.Variable;
 import org.springframework.data.mongodb.datatables.mapping.DataTablesInput;
@@ -14,9 +13,9 @@ import org.springframework.data.mongodb.datatables.mapping.DataTablesOutput;
 public interface VariableService {
     DataTablesOutput data(DataTablesInput input);
     
-    void insert(VariableInsertFilter filter, CheckError checkError);
+    void insert(VariableFilter filter, CheckError checkError);
     
-    void update(VariableUpdateFilter filter, CheckError checkError);
+    void update(VariableFilter filter, CheckError checkError);
     void delete(String... id);
     
     Variable findById(String id);